+++ river_route/configs/__init__.py
@@ -1,3 +1,3 @@
-from .Configs import Configs
+from .Configs import Configs, is_dev_null
 
-__all__ = ['Configs']
+__all__ = ['Configs', 'is_dev_null']